# server/go/model_authentication_request.go
/*
 * ECE 461 - Fall 2021 - Project 2
 *
 * API for ECE 461/Fall 2021/Project 2: A Trustworthy Module Registry
 *
 * API version: 2.0.0
 * Contact: davisjam@purdue.edu
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type AuthenticationRequest struct {

	User *User `json:"User"`

	Secret *UserAuthenticationInfo `json:"Secret"`
}