# server/go/model_package_history_entry.go
/*
 * ECE 461 - Fall 2021 - Project 2
 *
 * API for ECE 461/Fall 2021/Project 2: A Trustworthy Module Registry
 *
 * API version: 2.0.0
 * Contact: davisjam@purdue.edu
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

// One entry of the history of this package.
type PackageHistoryEntry struct {

	User *User `json:"User"`
	// Date of activity.
	Date time.Time `json:"Date"`

	PackageMetadata *PackageMetadata `json:"PackageMetadata"`

	Action string `json:"Action"`
}