# server/go/model_package_rating.go
/*
 * ECE 461 - Fall 2021 - Project 2
 *
 * API for ECE 461/Fall 2021/Project 2: A Trustworthy Module Registry
 *
 * API version: 2.0.0
 * Contact: davisjam@purdue.edu
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// Package rating (cf. Project 1).  If the Project 1 that you inherited does not support one or more of the original properties, denote this with the value \"-1\".
type PackageRating struct {

	BusFactor float64 `json:"BusFactor"`

	Correctness float64 `json:"Correctness"`

	RampUp float64 `json:"RampUp"`

	ResponsiveMaintainer float64 `json:"ResponsiveMaintainer"`

	LicenseScore float64 `json:"LicenseScore"`
	// The fraction of its dependencies that are pinned to at least a specific major+minor version, e.g. version 2.3.X of a package. (If there are zero dependencies, they should receive a 1.0 rating. If there are two dependencies, one pinned to this degree, then they should receive a Â½ = 0.5 rating).
	GoodPinningPractice float64 `json:"GoodPinningPractice"`
}